Snap: Simulate back button press when menu back button is pressed
Launching Gallery when the home button (shown as back button) is
pressed makes no sense, especially because it's only shown in Snap's
own gallery. In addition to that, if Gallery is disabled, Snap closes
itself. Fix both the issues by simply simulating a back button press
when the menu back button is pressed.

REGRESSION-607

Change-Id: I0b2b6ee074715e2866bcebb009118b8271b1efdf